% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{impute_matrix}
\alias{impute_matrix}
\title{Impute a matrix into another}
\usage{
impute_matrix(x, y, rows, cols = NULL, replace = FALSE)
}
\arguments{
\item{x}{the parent matrix which will be imputed. Note, this matrix must be at least the same
dimensions as \code{y}}

\item{y}{the matrix that will be imputed into \code{x}}

\item{rows}{the rows in \code{x} where the rows of \code{y} will be imputed. Note that it is assumed that
the rows in \code{y} are ordered so that they will go into \code{x} as specified by the rows argument}

\item{cols}{optional argument used when not dealing with a square \code{y} matrix in which the columns
and rows for imputation are the same.Defaults to \code{NULL}.  When imputing finite element stiffness
matrices, this will usually be 'NULL'  Functions the same as \code{rows} but for columns.}

\item{replace}{\code{boolean} for whether \code{y} will replace \code{x} at these positions or be added to \code{x}.
Typically in FEA, added is usual so the default value of \code{FALSE} would be used in this case.}
}
\value{
a matrix with the same dimensions as \code{x}
}
\description{
Imputes a matrix into another at the specified rows and columns
}
