% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{Ryacas_fns}
\alias{Ryacas_fns}
\title{\code{Ryacas} function names}
\format{
A vector of \code{Ryacas} function names which are used internally
so that \code{sub_eq} does not try to remove this part of the string when
performing the substitution.
}
\source{
lsf.str("package:Ryacas")
}
\usage{
Ryacas_fns
}
\description{
\code{Ryacas} function names
}
\keyword{datasets}
