% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{simplify_ryacas}
\alias{simplify_ryacas}
\title{Simplify \code{Ryacas} expression involving trigonometric operators}
\usage{
simplify_ryacas(x, type = "fraction", digits = 3)
}
\arguments{
\item{x}{a \code{Ryacas} symbolic expression}

\item{type}{choose either \code{fraction} for fractional form or
\code{round} to round to \code{digits}}

\item{digits}{required when using \code{round}}
}
\value{
a \code{Ryacas} symbolic expression
}
\description{
Simplify \code{Ryacas} expression involving trigonometric operators
}
