% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{simplify_trig}
\alias{simplify_trig}
\title{Simplify trigonometric \code{ysym} objects}
\usage{
simplify_trig(x, type = "fraction", ...)
}
\arguments{
\item{x}{the \code{ysym} object}

\item{type}{if \code{fraction} then will return the fractional result. If 'round', will
return the decimal result}
}
\value{
a ysym object
}
\description{
Simplify \code{ysym} objects with trig functions to either decimal or fractions
}
