% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{string_fraction}
\alias{string_fraction}
\title{Fractions of numbers in a string}
\usage{
string_fraction(x)
}
\arguments{
\item{x}{the string or vector of strings}
}
\value{
an object the same length as \code{x} but with all numbers turned to fractions
}
\description{
Uses \code{MASS::fractions} to turn decimal numbers in a string into fractions.
}
