% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fns.R
\name{string_round}
\alias{string_round}
\title{Round numbers in a string}
\usage{
string_round(x, dp = 3)
}
\arguments{
\item{x}{a string of vector of strings which may have a number within}

\item{dp}{digits to round to}
}
\value{
an object the same length as \code{x} but with all numbers rounded as per \code{dp}
}
\description{
Vectorised function that takes a string (or vector of) and rounds any numbers in the
string as dictated by \code{dp}.
}
