% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_sub.R
\name{sub_eq}
\alias{sub_eq}
\title{Substitute variables in equations}
\usage{
sub_eq(eq, val, vars = NULL, ...)
}
\arguments{
\item{eq}{the string representation of the equation to solve}

\item{val}{a vector where each value has a corresponding name
of a variable which appears in the equation to be solved}

\item{vars}{a list which the user can use to input the parts
of the equation}

\item{...}{similar to \code{vars} but does not require the user to
input the variables in list form}
}
\value{
a \code{list} including the following:
\enumerate{
\item The equation after multiplication and before substitution
\item The equation after multiplication and after substitution
\item The final solution
}
}
\description{
Allows the substitution but not solution of \code{Ryacas} equations
also allowing special variable names as found in engineering
mathematics such as \verb{theta_\{x, z\}}
}
\examples{
library(Ryacas)
library(magrittr)

xyz <- c('x', 'y', 'z')

# stress tensor
sigma <-
  sapply(xyz, function(x)
    paste('theta_{', xyz, ',', x, '}', sep = ''))
areas <- c('A_1', 'A_2', 'A_3')

vals <- 1:12
names(vals) <- c(sigma, areas)

# calculate the effective force 
sub_eq('s * a', vals, s = ysym(sigma), a = ysym(areas))

}
