% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_sub.R
\name{tex_sub}
\alias{tex_sub}
\title{Substitute variables in equation but don't solve}
\usage{
tex_sub(x, ...)
}
\arguments{
\item{x}{the input \code{TeX} string as produced by the \code{TeX} function in the
\code{package:Ryacas} package}

\item{...}{the names and values of variables for substitution i.e. \code{a = 4} will
search the string for \code{a} and substitute it for \code{4} but not simplify the output}
}
\description{
Given an equation in \code{TeX} form as produced by the \code{package:Ryacas} package,
the function will substitute variables for numerical values, but not solve
the equation.
}
\examples{
eq <- Ryacas::ysym(('(a * b) / c'))
tex_sub(Ryacas::tex(eq), a = 4, b = 3, c = 2)
}
