% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{YACAS_cmds}
\alias{YACAS_cmds}
\title{YACAS command names}
\format{
A vector of YACAS command names which should not be used
for variable names in user experessions - only to be used in
the user expresssions as yacas commands.
}
\source{
\url{http://yacas.sourceforge.net/ref.book.pdf}
}
\usage{
YACAS_cmds
}
\description{
YACAS command names
}
\keyword{datasets}
