% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{YACAS_fn_names}
\alias{YACAS_fn_names}
\title{YACAS command names}
\format{
A vector of YACAS command names:
}
\source{
\url{http://yacas.sourceforge.net/ref.book.pdf}
}
\usage{
YACAS_fn_names
}
\description{
YACAS command names
}
\keyword{datasets}
